# docs/swagger.yaml
basePath: /
definitions:
...
  fortyfour-backend_internal_dto.ErrorResponse:
    properties:
      message:
        example: invalid credentials
        type: string
    type: object
...
  fortyfour-backend_internal_dto.NotificationListResponse:
    properties:
      notifications:
        items:
          $ref: '#/definitions/fortyfour-backend_internal_dto.NotificationResponse'
        type: array
      unread_count:
        type: integer
    type: object
  fortyfour-backend_internal_dto.NotificationResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      read:
        type: boolean
      type:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API documentation for Fortyfour Backend - main auth and management
    service.
  title: Fortyfour Backend API
  version: "1.0"
paths:
...
  /api/notifications:
    delete:
      description: Menghapus semua notifikasi milik user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete all notifications
      tags:
      - Notifications
    get:
      description: Mengambil semua notifikasi milik user yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.NotificationListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all notifications
      tags:
      - Notifications
  /api/notifications/{id}:
    delete:
      description: Menghapus satu notifikasi berdasarkan ID
      parameters:
      - description: Notification ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a notification
      tags:
      - Notifications
  /api/notifications/{id}/read:
    patch:
      description: Menandai satu notifikasi sebagai sudah dibaca
      parameters:
      - description: Notification ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - Notifications
  /api/notifications/read-all:
    patch:
      description: Menandai semua notifikasi user sebagai sudah dibaca
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"